muppy/muppy.py
import gc
import sys

def get_objects():
    """Return a list of all known objects. """
    return gc.get_objects()

def sort(objects):
    """Sort objects by size in bytes."""
    objects.sort(lambda x, y: sys.getsizeof(x) - sys.getsizeof(y))
    return objects
    
def filter(objects, Type=object, min=-1, max=-1):
    """Filter objects."""
    res = []
    if min > max:
        raise ValueError("minimum must be smaller than maximum")
    [res.append(o) for o in objects if type(o) == Type]
    if min > -1:
        [res.remove(o) for o in res if sys.getsizeof(o) < min]
    if max > -1:
        [res.append(o) for o in res if sys.getsizeof(o) > max]
    return res

def get_referents(object, level=1):
    """Get all referents of an object up to a certain level.
    
    This function is recursive."""

    res = gc.get_referents(object)
    level -= 1
    if level > 0:
        for o in res:
            res.extend(get_referents(o, level))
    return res

def _remove_duplicates(objects):
    return list(set(objects))
    
